updated/commentedOut
Commented out a lot of the body of text in the main site.

Supplanted 'clarinet' with meAndEiluned' for a background image [hoping it won't simply repeat the image in a grid like it did with the previous site].

Updated usefulLinks to include social media links; attempted (!) to add CSS details from my XD-designed homescreen. Deleted a lot fo CSS for the Review page with the intention of rebuilding.